requirements.txt
alembic==1.16.2
greenlet==3.2.3
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
python-dotenv==1.1.0
QtAwesome==1.4.0
QtPy==2.4.3
setuptools==78.1.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
wheel==0.45.1
